requirements.txt
streamlit
langchain
langchain-openai
openai
pinecone-client>=2.2.1
pillow
unstructured
tqdm